requirements.txt
duckdb>=0.10
pandas
requests
tqdm
python-dateutil
